Created damage system

+++ Elementals/Assets/Scripts/Damage/DamageInfo.cs
@@ -2,7 +2,7 @@
 {
     public struct DamageInfo
     {
-        public int RawAmount { get; set; }
+        public float RawAmount { get; set; }
         public Element Element { get; set; }
     }
 }
